Version 1.0.0 patch; see notes.
This patch finally tags a version 1.0.0; I finally feel the code is good
enough--and I've made enough cleanup--to be confident in callin git 1.0.
Also, in addition to the version bump, this patch adds some syntax
cleanup, file renaming, makes static vs. shared library building an
option in CMake and adds a rounded_rectangle_center function that draws
the shape from the center out. Finally, there is a new cairockscffi
method added to the Python module called "cr_cast" that will try and
convert whatever is passed to it into a cairo_t*, in addition to finally
adding version information to Python.

There will CERTAINLY be bugs in this tag, but this code is getting
quite helpful.