Apply major pre-1.1.0 features; see notes
This is a rather large, unwieldy patch that should have been avoided if
I had managed my changes better. Regardless, I'm going to stash
everything here now and proceed in a more managed fashion moving
forward. NOTE: the incoming changes for both 1.1.0 and 1.2.0 are
currently in--and will be kept up to date in--the README file.